добавил класс Parser, переименовал файл, сделал функцию staticmethod, убрал лишний код
